# src/features/evidence/measures/dtos/request/reject-measure.request.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsOptional, IsString, MaxLength } from 'class-validator';

export class RejectMeasureRequestDto {
  @ApiProperty({
    example: 'duplicate',
    description: 'Why this proposed measure was rejected.',
    required: false,
  })
  @IsOptional()
  @IsString()
  @MaxLength(500)
  reason?: string;
}